requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
configargparse==0.15.1
idna==2.8                 # via requests
oauthlib==3.1.0           # via requests-oauthlib
pbr==5.4.3                # via stevedore
py-trello==0.15.0
python-dateutil==2.8.1    # via py-trello
pytz==2019.3              # via py-trello
pyyaml==5.1.2
requests-oauthlib==1.3.0  # via py-trello
requests==2.22.0          # via py-trello, requests-oauthlib
six==1.13.0               # via python-dateutil, stevedore
stevedore==1.31.0
urllib3==1.25.6           # via requests
